fix: Cast score to NUMERIC in search_unified to match return type
The function declares score NUMERIC in its return table, but CASE
expressions return plain integers. PostgreSQL rejects the type mismatch
at runtime, breaking all search functionality. Cast rr.score::NUMERIC
in the final SELECT.

+++ sql/schema/sot/MIG_792__search_unified_add_coordinates.sql
@@ -244,7 +244,7 @@ BEGIN
             ELSE 'weak'
         END AS match_strength,
         rr.match_reason,
-        rr.score,
+        rr.score::NUMERIC,
         rr.metadata
     FROM ranked_results rr
     WHERE rr.score > 0
